requirements.txt
# Core dependencies
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
Pillow>=10.0.0
pandas>=2.0.0

# AI/ML Utilities
google-generativeai>=0.3.0
tqdm>=4.66.0

# Utilities
requests>=2.31.0
PyYAML>=6.0
